feat: add options with escapeSpecialCharacters (#65)
Fixes #26. Fixes #63.

As described in #63, this introduces the concept of options with single
option, `escapeSpecialCharacters`. The option defaults to:

* `true` when called for a template literal's string tag
* `false` when called as a function

```js
// "\$hello!"
dedent.options({ escapeSpecialCharacters: false })`
  $hello!
`;
```

I'd played with allowing passing it in as a first argument instead of a
string or template literal strings array, but that got complex. I
suppose we can do that as a followup if people really want.

cc @G-Rath @sirian - what do you think?